# Conflicts:
#	trunk/SUAVE/Analyses/Aerodynamics/Results.py
#	trunk/SUAVE/Components/Energy/Networks/Turbofan.py
#	trunk/SUAVE/Methods/Aerodynamics/AERODAS/post_stall_coefficients.py
#	trunk/SUAVE/Methods/Aerodynamics/AERODAS/pre_stall_coefficients.py
#
trunk/SUAVE/Methods/Aerodynamics/Common/Fidelity_Zero/Drag/compressibili
ty_drag_wing_total.py
#
trunk/SUAVE/Methods/Aerodynamics/Common/Fidelity_Zero/Drag/parasite_tota
l.py
#
trunk/SUAVE/Methods/Aerodynamics/Common/Fidelity_Zero/Helper_Functions/w
ave_drag_lift.py
#
trunk/SUAVE/Methods/Aerodynamics/Fidelity_Zero/Drag/miscellaneous_drag_a
ircraft.py
#
trunk/SUAVE/Methods/Aerodynamics/Fidelity_Zero/Drag/parasite_drag_fusela
ge.py
#
trunk/SUAVE/Methods/Aerodynamics/Fidelity_Zero/Lift/linear_inviscid_wing
.py
#
trunk/SUAVE/Methods/Aerodynamics/Fidelity_Zero/Lift/linear_supersonic_li
ft.py
#	trunk/SUAVE/Methods/Aerodynamics/Fidelity_Zero/Lift/vortex_lift.py
#	trunk/SUAVE/Methods/Aerodynamics/OpenVSP_Wave_Drag/wave_drag_lift.py
#	trunk/SUAVE/Methods/Aerodynamics/SU2_Euler/untrimmed.py
#
trunk/SUAVE/Methods/Aerodynamics/Supersonic_Zero/Drag/compressible_mixed
_flat_plate.py
#
trunk/SUAVE/Methods/Aerodynamics/Supersonic_Zero/Drag/compressible_turbu
lent_flat_plate.py
#
trunk/SUAVE/Methods/Aerodynamics/Supersonic_Zero/Drag/parasite_drag_airc
raft.py
#
trunk/SUAVE/Methods/Aerodynamics/Supersonic_Zero/Drag/parasite_drag_wing
.py
#
trunk/SUAVE/Methods/Aerodynamics/Supersonic_Zero/Drag/parasite_total.py

+++ trunk/SUAVE/Methods/Missions/Segments/Climb/Constant_Speed_Constant_Angle_Noise.py
@@ -47,7 +47,7 @@ def expand_state(segment,state):
     
     #number of time steps (space discretization)
     total_time=(x0+500)/v_x    
-    n_points   = np.ceil(total_time/dt +1)       
+    n_points   = np.int(np.ceil(total_time/dt +1))       
     
     state.numerics.number_control_points = n_points
